requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
pypdf
chromadb
sentence-transformers
python-dotenv
langchain-chroma
ipykernel
fastapi
python-multipart
uvicorn
langchain-experimental
python-docx
requests
langchain-google-genai
pymupdf
unstructured
unstructured[pdf]
langchain-ollama